Refactor enable-block-product-editor
